+++ front/components/agent_builder/capabilities/AgentBuilderCapabilitiesBlock.tsx
@@ -1,15 +1,12 @@
 import {
   BookOpenIcon,
   Button,
-  Card,
-  CardActionButton,
   CardGrid,
   ContentMessage,
   EmptyCTA,
   Hoverable,
   Spinner,
   ToolsIcon,
-  XMarkIcon,
 } from "@dust-tt/sparkle";
 import React, { useMemo, useState } from "react";
 import { useFieldArray, useFormContext } from "react-hook-form";
@@ -25,90 +22,11 @@ import { getSpaceIdToActionsMap } from "@app/components/agent_builder/get_spacei
 import { useMCPServerViewsContext } from "@app/components/agent_builder/MCPServerViewsContext";
 import { useSpacesContext } from "@app/components/agent_builder/SpacesContext";
 import type { AgentBuilderAction } from "@app/components/agent_builder/types";
-import {
-  getDefaultMCPAction,
-  isDefaultActionName,
-} from "@app/components/agent_builder/types";
+import { getDefaultMCPAction } from "@app/components/agent_builder/types";
+import { ActionCard } from "@app/components/shared/tools_picker/ActionCard";
 import { BACKGROUND_IMAGE_STYLE_PROPS } from "@app/components/shared/tools_picker/util";
-import { getMcpServerViewDisplayName } from "@app/lib/actions/mcp_helper";
-import { getAvatar } from "@app/lib/actions/mcp_icons";
-import { MCP_SPECIFICATION } from "@app/lib/actions/utils";
-import type { MCPServerViewType } from "@app/lib/api/mcp";
 import type { TemplateActionPreset } from "@app/types";
 import { pluralize } from "@app/types";
-
-function actionIcon(
-  action: AgentBuilderAction,
-  mcpServerView: MCPServerViewType | null
-) {
-  if (mcpServerView?.server) {
-    return getAvatar(mcpServerView.server, "xs");
-  }
-}
-
-function actionDisplayName(
-  action: AgentBuilderAction,
-  mcpServerView: MCPServerViewType | null
-) {
-  if (mcpServerView && action.type === "MCP") {
-    return getMcpServerViewDisplayName(mcpServerView, action);
-  }
-
-  return `${MCP_SPECIFICATION.label}${
-    !isDefaultActionName(action) ? " - " + action.name : ""
-  }`;
-}
-
-export interface ActionCardProps {
-  action: AgentBuilderAction;
-  onRemove: () => void;
-  onEdit?: () => void;
-}
-
-export function ActionCard({ action, onRemove, onEdit }: ActionCardProps) {
-  const { mcpServerViews, isMCPServerViewsLoading } =
-    useMCPServerViewsContext();
-
-  const mcpServerView =
-    action.type === "MCP" && !isMCPServerViewsLoading
-      ? (mcpServerViews.find(
-          (mcpServerView) =>
-            mcpServerView.sId === action.configuration.mcpServerViewId
-        ) ?? null)
-      : null;
-
-  const displayName = actionDisplayName(action, mcpServerView);
-  const description = action.description ?? "";
-
-  return (
-    <Card
-      variant="primary"
-      className="h-28"
-      onClick={onEdit}
-      action={
-        <CardActionButton
-          size="mini"
-          icon={XMarkIcon}
-          onClick={(e: Event) => {
-            onRemove();
-            e.stopPropagation();
-          }}
-        />
-      }
-    >
-      <div className="flex w-full flex-col gap-2 text-sm">
-        <div className="flex w-full items-center gap-2 font-medium text-foreground dark:text-foreground-night">
-          {actionIcon(action, mcpServerView)}
-          <span className="truncate">{displayName}</span>
-        </div>
-
-        <div className="text-muted-foreground dark:text-muted-foreground-night">
-          <span className="line-clamp-2 break-words">{description}</span>
-        </div>
-      </div>
-    </Card>
-  );
-}
 
 interface AgentBuilderCapabilitiesBlockProps {
   isActionsLoading: boolean;
