+++ front/components/shared/tools_picker/ActionCard.tsx
@@ -1,0 +1,82 @@
+import { Card, CardActionButton, XMarkIcon } from "@dust-tt/sparkle";
+
+import { useMCPServerViewsContext } from "@app/components/agent_builder/MCPServerViewsContext";
+import type { AgentBuilderAction } from "@app/components/agent_builder/types";
+import { isDefaultActionName } from "@app/components/agent_builder/types";
+import { getMcpServerViewDisplayName } from "@app/lib/actions/mcp_helper";
+import { getAvatar } from "@app/lib/actions/mcp_icons";
+import { MCP_SPECIFICATION } from "@app/lib/actions/utils";
+import type { MCPServerViewType } from "@app/lib/api/mcp";
+
+function actionIcon(
+  action: AgentBuilderAction,
+  mcpServerView: MCPServerViewType | null
+) {
+  if (mcpServerView?.server) {
+    return getAvatar(mcpServerView.server, "xs");
+  }
+}
+
+function actionDisplayName(
+  action: AgentBuilderAction,
+  mcpServerView: MCPServerViewType | null
+) {
+  if (mcpServerView && action.type === "MCP") {
+    return getMcpServerViewDisplayName(mcpServerView, action);
+  }
+
+  return `${MCP_SPECIFICATION.label}${
+    !isDefaultActionName(action) ? " - " + action.name : ""
+  }`;
+}
+
+export interface ActionCardProps {
+  action: AgentBuilderAction;
+  onRemove: () => void;
+  onEdit?: () => void;
+}
+
+export function ActionCard({ action, onRemove, onEdit }: ActionCardProps) {
+  const { mcpServerViews, isMCPServerViewsLoading } =
+    useMCPServerViewsContext();
+
+  const mcpServerView =
+    action.type === "MCP" && !isMCPServerViewsLoading
+      ? (mcpServerViews.find(
+          (mcpServerView) =>
+            mcpServerView.sId === action.configuration.mcpServerViewId
+        ) ?? null)
+      : null;
+
+  const displayName = actionDisplayName(action, mcpServerView);
+  const description = action.description ?? "";
+
+  return (
+    <Card
+      variant="primary"
+      className="h-28"
+      onClick={onEdit}
+      action={
+        <CardActionButton
+          size="mini"
+          icon={XMarkIcon}
+          onClick={(e: Event) => {
+            onRemove();
+            e.stopPropagation();
+          }}
+        />
+      }
+    >
+      <div className="flex w-full flex-col gap-2 text-sm">
+        <div className="flex w-full items-center gap-2 font-medium text-foreground dark:text-foreground-night">
+          {actionIcon(action, mcpServerView)}
+          <span className="truncate">{displayName}</span>
+        </div>
+
+        <div className="text-muted-foreground dark:text-muted-foreground-night">
+          <span className="line-clamp-2 break-words">{description}</span>
+        </div>
+      </div>
+    </Card>
+  );
+}
